description
This is a small example of 'keyword-search' about 'react'. throught
'ref' attribute access to the user manual in input the content of the
input in the input box,Through 'ajax' sends a request to the backstage,
through fuzzy select, data from webpage in the string is returned to
webpage contain the keywords, the user can through the mouse to choose,
can also use the keyboard up and down keys to select, move up and down,
by giving the current this. When the 'this.index' of 'li' added The
'active' class, in order to achieve the effect of selected line is
highlighted.